Added Error property to AccountInfoResponse
It is possible that an account has not been used when an AccountInfoRequest is sent for that account. In that case, an "Error" field is returned by the server. This change allows that field to be populated. This is useful for detecting accounts that have not had their first block created.